requirements.txt
flask
nomic
pytest
pyllamacpp
pyyaml
grpcio
grpcio-tools
futures